Add client files, try to get userId
