+++ client/src/pages/Search/Search.js
@@ -15,12 +15,12 @@ import "../Search/search.css";
 
 const Search = () => {
     const [salons, setSalons] = useState(null);
-    console.log(salons);
+    // console.log(salons);
     const [selectedLocation, setSelectedLocation] = useState("");
     const [selectedServices, setSelectedServices] = useState([]);
 
     useEffect(() => {
-        console.log("hello")
+        // console.log("hello")
         async function getSalons() {
             if (!salons) {
                 const response = await getAllSalons();
@@ -44,9 +44,10 @@ const Search = () => {
         filter();
     }, [selectedLocation, selectedServices]);
 
-    // const userId = 
+    // const userId = props.userId or select user id redux
 
     const getSalonIsFavourite = (salonId) => {
+        console.log(salonId);
      // check if there is entry in favourites table that has salonId and UserId
      // need to get logged in user as a prop from home - will be an user object from DB
        
@@ -58,11 +59,8 @@ const Search = () => {
 
     };
 
-    // user id will be a global variable that i'll fetch from the DB on page load - logic will be done on Homepage
-
 
     const setLocation = (event) => {
-        console.log("hi");
         console.log(event);
         const location = event.value
         setSelectedLocation(location);
@@ -113,8 +111,7 @@ const Search = () => {
                                 twitterLink={salon.twitter}
                                 salonIsFavourite={getSalonIsFavourite(salon._id)}
                                 salonId={salon._id}
-                    
-
+                                key={salon._id}
                             />)
                     ) : (
                         <p>No salons found</p>
